Set expiration per session.
In some cases it might be legitimate to override the session expiration. For
instance, in an authentication workflow, one might want to check the
"Remember me" option and in that case the session cookie should not expire
anymore while by default all new cookies should inherit from the default
configuration.

This change will allow a user to call SetExpiration(...) on a storer
so that the new expiration is applied to that session. This updates the cookie
with the new expiration. Please note that the regenerating a session will reset
the expiration to the default value which was globally configured.