Add UpdateExpires method in order to override cookie expiration.
In some cases it might be legitimate to override the cookie expiration. For
instance, in an authentication workflow, one might want to check the
"Remember me" option and in that case the session cookie should not expire
anymore while by default all new cookies should inherit from the default
configuration.