Allow setting cookie secure flag when server is behing a proxy
If the communication between the proxy and fasthttp is not over TLS, for
instance when both are on the same host, the secure flag could still be
legitimately set given the proxy does the TLS termination.

IsSecureFunc configuration option is introduced to provide a function checking
whether the communication between the client and the server is secure.
In that case, the cookie secure flag should be set according to Secure
configuration option.

For supporting a server behind a proxy, one can override the default
IsSecureFunc function to check whether X-Forwarded-Proto header is set to
'https' and whether the remote client is a trusted proxy or not.